pyproject.toml
[project]
name = "echo"
version = "0.1.0"
description = "Django project with PostgreSQL integration"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.0.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
    "whitenoise[brotli]>=6.11.0",
]
